Add IntegrationMethod class
This class provides a means to specify the quadrature used in
SurfaceExtraction::integrate for the integration over u and v. Some
commonly used methods (Trapezium, Midpoint, Simpson's and Boole's) have
been statically defined.

Since Simpson's rule has been set as the default method for
SphericalExtraction::integrate_mode, the parameter
num_points_theta is now required to be odd. If an even value is set, it
is automatically incremented by 1. The SphericalGeometry class has also
been modified to include the poles (even though they do not contribute
to the integrals since the area element vanishes there) to reflect the
change in integration method.